Ajout d'une search bar et affichage des pages en fonction du carnet correspondant
