+++ BagManIOs/View/PageView.swift
@@ -2,125 +2,14 @@
 //  PageView.swift
 //  BagManIOs
 //
-//  Created by Sabri on 21/12/2017.
+//  Created by Sabri on 22/12/2017.
 //  Copyright © 2017 utt. All rights reserved.
 //
 
 import UIKit
 import SQLite
 
-class PageView: UITableViewController {
-    var page: [Page] = []
-    var carnet_id: Int!
-    var idCarnetCell: String = "cell"
-    var database: Connection!
-    var db: Bdd!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        self.db = Bdd()
-        //carnet_id = 1
-        page = db.getPagesByCarnet(carnetId_pf: carnet_id)
-        print(page)
-        // Uncomment the following line to preserve selection between presentations
-        // self.clearsSelectionOnViewWillAppear = false
-        
-//         Uncomment the following line to display an Edit button in the navigation bar for this view controller.
-//         self.navigationItem.rightBarButtonItem = self.editButtonItem
-    }
-    
-//    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-//        <#code#>
-//    }
-    
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-    // MARK: - Table view data source
-    
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        // #warning Incomplete implementation, return the number of sections
-        return 1
-    }
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return page.count
-    }
-    
-    override func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        // #warning Incomplete implementation, return the number of rows
-        return "Liste des pages"
-    }
-    
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: idCarnetCell, for: indexPath)
-        
-        cell.textLabel?.text = "\(page[indexPath.row].title)"
-        print(indexPath.section)
-        print(indexPath.row)
-        cell.detailTextLabel?.text = "Dernière mis à jour le : \(page[indexPath.row].updatedAt)"
-        //        return cell
-        
-        return cell
-    }
-    
-    //    func remplirTableau() {
-    //        for c in carnet {
-    //            carnet.append(c)
-    //        }
-    //    }
-    
-    
-    /*
-     // Override to support conditional editing of the table view.
-     override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-     // Return false if you do not want the specified item to be editable.
-     return true
-     }
-     */
-    
-    /*
-     // Override to support editing the table view.
-     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-     if editingStyle == .delete {
-     // Delete the row from the data source
-     tableView.deleteRows(at: [indexPath], with: .fade)
-     } else if editingStyle == .insert {
-     // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-     }
-     }
-     */
-    
-    /*
-     // Override to support rearranging the table view.
-     override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
-     
-     }
-     */
-    
-    /*
-     // Override to support conditional rearranging of the table view.
-     override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-     // Return false if you do not want the item to be re-orderable.
-     return true
-     }
-     */
-    
-    /*
-     // MARK: - Navigation
-     
-     // In a storyboard-based application, you will often want to do a little preparation before navigation
-     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-     // Get the new view controller using segue.destinationViewController.
-     // Pass the selected object to the new view controller.
-     }
-     */
-    
-}
-
+//class PageView: UIViewController {
+//    <#code#>
+//}
